requirements.txt
# Dependências para Istio Templates Renderer
# Apenas o essencial para o helm_render.py

jinja2>=3.1.0,<4.0
pyyaml>=6.0,<7.0
azure-identity>=1.17.1,<2.0
azure-keyvault-secrets>=4.8.0,<5.0
